+++ module/item/entity.js
@@ -7,6 +7,7 @@ import { ChatAttack } from "../misc/chat-attack.js";
 import { SemanticVersion } from "../semver.js";
 import { ItemChange } from "./components/change.js";
 import { getHighestChanges } from "../actor/apply-changes.js";
+import { RollPF } from "../roll.js";
 
 /**
  * Override and extend the basic :class:`Item` implementation
@@ -130,7 +131,7 @@ export class ItemPF extends Item {
 
     const formula = getProperty(this.data, "data.uses.autoDeductChargesCost");
     if (!(typeof formula === "string" && formula.length > 0)) return 1;
-    const cost = new Roll(formula, this.getRollData()).roll().total;
+    const cost = RollPF.safeRoll(formula, this.getRollData()).total;
     return cost;
   }
 
@@ -183,11 +184,11 @@ export class ItemPF extends Item {
       case "reach":
         return convertDistance(getProperty(this.getRollData(), "range.reach") || 0)[0];
       case "close":
-        return convertDistance(new Roll("25 + floor(@cl / 2) * 5", this.getRollData()).roll().total)[0];
+        return convertDistance(RollPF.safeRoll("25 + floor(@cl / 2) * 5", this.getRollData()).total)[0];
       case "medium":
-        return convertDistance(new Roll("100 + @cl * 10", this.getRollData()).roll().total)[0];
+        return convertDistance(RollPF.safeRoll("100 + @cl * 10", this.getRollData()).total)[0];
       case "long":
-        return convertDistance(new Roll("400 + @cl * 40", this.getRollData()).roll().total)[0];
+        return convertDistance(RollPF.safeRoll("400 + @cl * 40", this.getRollData()).total)[0];
       case "mi":
         return convertDistance(range * 5280)[0];
       default:
@@ -247,8 +248,8 @@ export class ItemPF extends Item {
       if (spellbook != null) {
         try {
           result =
-            new Roll(spellbook.baseDCFormula, rollData).roll().total +
-            new Roll(data.save.dc.length > 0 ? data.save.dc : "0", rollData).roll().total +
+            Roll.create(spellbook.baseDCFormula, rollData).roll().total +
+            Roll.create(data.save.dc.length > 0 ? data.save.dc : "0", rollData).roll().total +
             dcBonus;
         } catch (e) {
           console.error(e, spellbook.baseDCFormula, data.save.dc.length > 0 ? data.save.dc : "0");
@@ -258,7 +259,7 @@ export class ItemPF extends Item {
     }
     const dcFormula = getProperty(data, "save.dc")?.toString() || "0";
     try {
-      result = new Roll(dcFormula, rollData).roll().total + dcBonus;
+      result = Roll.create(dcFormula, rollData).roll().total + dcBonus;
     } catch (e) {
       console.error(e, dcFormula);
     }
@@ -593,7 +594,7 @@ export class ItemPF extends Item {
         rng.long = null;
       } else if (typeof rng.value === "string" && rng.value.length) {
         try {
-          rng.value = new Roll(rng.value, this.getRollData()).roll().total.toString();
+          rng.value = Roll.create(rng.value, this.getRollData()).roll().total.toString();
         } catch (err) {
           console.error(err);
         }
@@ -605,11 +606,7 @@ export class ItemPF extends Item {
       let dur = duplicate(data.duration || {});
       if (["inst", "perm", "spec", "seeText"].includes(dur.units)) dur.value = game.i18n.localize("PF1.Duration") + ":";
       else if (typeof dur.value === "string") {
-        try {
-          dur.value = new Roll(dur.value || "0", this.getRollData()).roll().total.toString();
-        } catch (err) {
-          console.error(this.name, "- Duration -", err);
-        }
+        dur.value = RollPF.safeRoll(dur.value || "0", this.getRollData(), [this.name, "Duration"]).total.toString();
       }
       labels.duration = [dur.value, C.timePeriods[dur.units]].filterJoin(" ");
     }
@@ -1124,7 +1121,7 @@ export class ItemPF extends Item {
     if (hasProperty(this.data, "data.uses.maxFormula")) {
       const maxFormula = getProperty(this.data, "data.uses.maxFormula");
       if (maxFormula !== "" && !formulaHasDice(maxFormula)) {
-        let roll = new Roll(maxFormula, rollData).roll();
+        let roll = RollPF.safeRoll(maxFormula, rollData);
         setProperty(this.data, "data.uses.max", roll.total);
       } else if (formulaHasDice(maxFormula)) {
         const msg = game.i18n
@@ -1219,7 +1216,7 @@ export class ItemPF extends Item {
         `data.attributes.spells.spellbooks.${this.data.data.spellbook}`
       );
       if (spellbook && spellbook.arcaneSpellFailure) {
-        templateData.spellFailure = new Roll("1d100").roll().total;
+        templateData.spellFailure = Roll.create("1d100").roll().total;
         templateData.spellFailureSuccess = templateData.spellFailure > this.parentActor.spellFailure;
       }
     }
@@ -1306,15 +1303,10 @@ export class ItemPF extends Item {
             break;
           case "ft":
           case "mi":
-            rangeValue = convertDistance(
-              new Roll(data.range.value.length > 0 ? data.range.value : "0", rollData).roll().total,
-              data.range.units
-            );
+            rangeValue = convertDistance(RollPF.safeRoll(data.range.value || "0", rollData).total, data.range.units);
             break;
           case "spec":
-            rangeValue = convertDistance(
-              new Roll(data.range.value.length > 0 ? data.range.value : "0", rollData).roll().total
-            );
+            rangeValue = convertDistance(RollPF.safeRoll(data.range.value || "0", rollData).total);
             break;
         }
         dynamicLabels.range =
@@ -1325,7 +1317,7 @@ export class ItemPF extends Item {
       // Duration
       if (data.duration != null) {
         if (!["inst", "perm"].includes(data.duration.units) && typeof data.duration.value === "string") {
-          let duration = new Roll(data.duration.value.length > 0 ? data.duration.value : "0", rollData).roll().total;
+          let duration = RollPF.safeRoll(data.duration.value || "0", rollData).total;
           dynamicLabels.duration = [duration, CONFIG.PF1.timePeriods[data.duration.units]].filterJoin(" ");
         }
       }
@@ -1485,16 +1477,16 @@ export class ItemPF extends Item {
 
   parseFormulaicAttacks({ formula = null } = {}) {
     const exAtkCountFormula = formula ?? (this.data.data.formulaicAttacks?.count?.formula || "");
-    let extraAttacks = 0;
+    let extraAttacks = 0,
+      xaroll;
     const rollData = this.getRollData();
-    try {
-      if (exAtkCountFormula.length > 0) {
-        const xaroll = new Roll(exAtkCountFormula, rollData).roll();
-        extraAttacks = Math.min(50, Math.max(0, xaroll.total)); // Arbitrarily clamp attacks
-      }
-    } catch (err) {
+    if (exAtkCountFormula.length > 0) {
+      xaroll = RollPF.safeRoll(exAtkCountFormula, rollData);
+      extraAttacks = Math.min(50, Math.max(0, xaroll.total)); // Arbitrarily clamp attacks
+    }
+    if (xaroll?.err) {
       const msg = game.i18n.localize("PF1.ErrorItemFormula").format(this.name, this.actor?.name);
-      console.warn(msg, err, exAtkCountFormula);
+      console.warn(msg, xaroll.err, exAtkCountFormula);
       ui.notifications.warn(msg);
     }
 
@@ -1503,7 +1495,7 @@ export class ItemPF extends Item {
     try {
       if (exAtkBonusFormula.length > 0) {
         rollData["attackCount"] = 1;
-        new Roll(exAtkBonusFormula, rollData).roll().total;
+        Roll.create(exAtkBonusFormula, rollData).roll().total;
       }
     } catch (err) {
       const msg = game.i18n.localize("PF1.ErrorItemFormula").format(this.name, this.actor?.name);
@@ -1697,7 +1689,7 @@ export class ItemPF extends Item {
             const fatlabel = this.data.data.formulaicAttacks.label || game.i18n.localize("PF1.FormulaAttack");
             for (let i = 0; i < exAtkCount; i++) {
               frollData["formulaicAttack"] = i + 1; // Add and update attack counter
-              const bonus = new Roll(exAtkBonusFormula, frollData).roll().total;
+              const bonus = RollPF.safeRoll(exAtkBonusFormula, frollData).total;
               allAttacks.push({
                 bonus: bonus.toString(),
                 label: fatlabel.format(i + 2),
@@ -1736,16 +1728,14 @@ export class ItemPF extends Item {
           for (const [i, modifier] of conditional.modifiers.entries()) {
             // Adds a formula's result to rollData to allow referencing it.
             // Due to being its own roll, this will only correctly work for static formulae.
-            // In try-block to avoid stalling due to malformed modifier
-            try {
-              conditionalData[[tag, i].join(".")] = new Roll(modifier.formula, rollData).roll().total;
-            } catch (e) {
+            const conditionalRoll = RollPF.safeRoll(modifier.formula, rollData);
+            if (conditionalRoll.err) {
               const msg = game.i18n.format("PF1.WarningConditionalRoll", { number: i + 1, name: conditional.name });
               console.warn(msg);
               ui.notifications.warn(msg);
               // Skip modifier to avoid multiple errors from one non-evaluating entry
               continue;
-            }
+            } else conditionalData[[tag, i].join(".")] = RollPF.safeRoll(modifier.formula, rollData).total;
 
             // Create a key string for the formula array
             const partString = `${modifier.target}.${modifier.subTarget}${
@@ -1773,21 +1763,17 @@ export class ItemPF extends Item {
         for (const target of ["effect.cl", "effect.dc", "misc.charges"]) {
           if (conditionalPartsCommon[target] != null) {
             const formula = conditionalPartsCommon[target].join("+");
-            try {
-              const roll = new Roll(formula, rollData).roll().total;
-              switch (target) {
-                case "effect.cl":
-                  rollData.cl += roll;
-                  break;
-                case "effect.dc":
-                  rollData.dcBonus = roll;
-                  break;
-                case "misc.charges":
-                  rollData.chargeCostBonus = roll;
-                  break;
-              }
-            } catch (e) {
-              console.error(e, formula);
+            const roll = RollPF.safeRoll(formula, rollData, [target, formula]).total;
+            switch (target) {
+              case "effect.cl":
+                rollData.cl += roll;
+                break;
+              case "effect.dc":
+                rollData.dcBonus = roll;
+                break;
+              case "misc.charges":
+                rollData.chargeCostBonus = roll;
+                break;
             }
           }
         }
@@ -2052,7 +2038,7 @@ export class ItemPF extends Item {
         // Determine size
         let dist = getProperty(this.data, "data.measureTemplate.size");
         if (typeof dist === "string") {
-          dist = new Roll(getProperty(this.data, "data.measureTemplate.size"), rollData).roll().total;
+          dist = RollPF.safeRoll(getProperty(this.data, "data.measureTemplate.size"), rollData).total;
         }
         dist = convertDistance(dist)[0];
 
@@ -2274,7 +2260,7 @@ export class ItemPF extends Item {
           if (range != null) {
             templateData.range = range;
             if (typeof range === "string") {
-              templateData.range = new Roll(range, rollData).roll().total;
+              templateData.range = RollPF.safeRoll(range, rollData).total;
               templateData.rangeFormula = range;
             }
             templateData.rangeLabel = `${templateData.range} ft.`;
@@ -2296,7 +2282,7 @@ export class ItemPF extends Item {
             `data.attributes.spells.spellbooks.${this.data.data.spellbook}`
           );
           if (spellbook && spellbook.arcaneSpellFailure) {
-            templateData.spellFailure = new Roll("1d100").roll().total;
+            templateData.spellFailure = Roll.create("1d100").roll().total;
             templateData.spellFailureSuccess = templateData.spellFailure > this.parentActor.spellFailure;
           }
         }
@@ -2520,7 +2506,7 @@ export class ItemPF extends Item {
     parts = parts.concat(extraParts);
     // Add attack bonus
     if (typeof itemData.attackBonus === "string" && itemData.attackBonus !== "") {
-      let attackBonus = new Roll(itemData.attackBonus, rollData).roll().total;
+      let attackBonus = RollPF.safeRoll(itemData.attackBonus, rollData).total;
       rollData.item.attackBonus = attackBonus;
       parts.push("@item.attackBonus");
     }
@@ -2582,7 +2568,7 @@ export class ItemPF extends Item {
     if (primaryAttack === false) parts.push("-5");
     // Add bonus
     if (bonus) {
-      rollData.bonus = new Roll(bonus, rollData).roll().total;
+      rollData.bonus = RollPF.safeRoll(bonus, rollData).total;
       parts.push("@bonus");
     }
 
@@ -2591,7 +2577,7 @@ export class ItemPF extends Item {
 
     if ((rollData.d20 ?? "") === "") rollData.d20 = "1d20";
 
-    let roll = new Roll([rollData.d20, ...parts].join("+"), rollData).roll();
+    let roll = RollPF.safeRoll([rollData.d20, ...parts].join("+"), rollData);
     return roll;
   }
 
@@ -2657,7 +2643,7 @@ export class ItemPF extends Item {
     rollData.item = itemData;
     const title = `${this.name} - ${game.i18n.localize("PF1.OtherFormula")}`;
 
-    const roll = new Roll(itemData.formula, rollData).roll();
+    const roll = RollPF.safeRoll(itemData.formula, rollData);
     return roll.toMessage({
       speaker: ChatMessage.getSpeaker({ actor: this.parentActor }),
       flavor: itemData.chatFlavor || title,
@@ -2764,7 +2750,7 @@ export class ItemPF extends Item {
       let rollParts = [];
       if (a === 0) rollParts = [...part.extra, ...extraParts];
       const roll = {
-        roll: new Roll([part.base, ...rollParts].join("+"), rollData).roll(),
+        roll: RollPF.safeRoll([part.base, ...rollParts].join("+"), rollData),
         damageType: part.damageType,
         type: part.type,
       };
@@ -2801,7 +2787,7 @@ export class ItemPF extends Item {
     });
     // Submit the roll to chat
     if (effectStr === "") {
-      new Roll(parts.join("+")).toMessage({
+      Roll.create(parts.join("+")).toMessage({
         speaker: ChatMessage.getSpeaker({ actor: this.parentActor }),
         flavor: game.i18n.localize("PF1.UsesItem").format(this.name),
       });
@@ -2809,7 +2795,7 @@ export class ItemPF extends Item {
       const chatTemplate = "systems/pf1/templates/chat/roll-ext.hbs";
       const chatTemplateData = { hasExtraText: true, extraText: effectStr };
       // Execute the roll
-      let roll = new Roll(parts.join("+"), data).roll();
+      let roll = RollPF.safeRoll(parts.join("+"), data);
 
       // Create roll template data
       const rollData = mergeObject(
@@ -3246,7 +3232,7 @@ export class ItemPF extends Item {
   getSpellPointCost(rollData = null) {
     if (!rollData) rollData = this.getRollData();
 
-    const roll = new Roll(getProperty(this.data, "data.spellPoints.cost") || "0", rollData).roll();
+    const roll = RollPF.safeRoll(getProperty(this.data, "data.spellPoints.cost") || "0", rollData);
     return roll.total;
   }
 
@@ -3344,15 +3330,15 @@ export class ItemPF extends Item {
     data.data.range.value = origData.data.range.value;
     switch (data.data.range.units) {
       case "close":
-        data.data.range.value = new Roll("25 + floor(@cl / 2) * 5", { cl: slcl[1] }).roll().total.toString();
+        data.data.range.value = RollPF.safeRoll("25 + floor(@cl / 2) * 5", { cl: slcl[1] }).total.toString();
         data.data.range.units = "ft";
         break;
       case "medium":
-        data.data.range.value = new Roll("100 + @cl * 10", { cl: slcl[1] }).roll().total.toString();
+        data.data.range.value = RollPF.safeRoll("100 + @cl * 10", { cl: slcl[1] }).total.toString();
         data.data.range.units = "ft";
         break;
       case "long":
-        data.data.range.value = new Roll("400 + @cl * 40", { cl: slcl[1] }).roll().total.toString();
+        data.data.range.value = RollPF.safeRoll("400 + @cl * 40", { cl: slcl[1] }).total.toString();
         data.data.range.units = "ft";
         break;
     }